compat+fix: add _strlwr/_strupr stubs; fix pointer-to-unsigned casts in wwstring
- windows.h: add _strlwr/_strupr inline stubs (MSVC string case-convert)
- windows.h: add _wcslwr/_wcsupr wide-char stubs
- wwstring.cpp: change reinterpret_cast<unsigned>(ptr) to
  reinterpret_cast<uintptr_t>(ptr) and local vars to uintptr_t so the
  pointer arithmetic is correct on 64-bit Clang/ARM64 (was a hard error)

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWLib/wwstring.cpp
@@ -102,9 +102,9 @@ StringClass::Get_String (int length, bool is_temp)
 				//
 				//	Grab this unused buffer for our string
 				//
-				unsigned temp_string=reinterpret_cast<unsigned>(m_TempStrings);
+				uintptr_t temp_string=reinterpret_cast<uintptr_t>(m_TempStrings);
 				temp_string+=MAX_TEMP_BYTES*MAX_TEMP_STRING;
-				temp_string&=~(MAX_TEMP_BYTES*MAX_TEMP_STRING-1);
+				temp_string&=~(uintptr_t)(MAX_TEMP_BYTES*MAX_TEMP_STRING-1);
 				temp_string+=index*MAX_TEMP_BYTES;
 				temp_string+=sizeof(_HEADER);	// The buffer contains header as well, and it needs to be at the start
 				string=reinterpret_cast<char*>(temp_string);
@@ -197,8 +197,8 @@ StringClass::Free_String (void)
 {
 	if (m_Buffer != m_EmptyString) {
 
-		unsigned buffer_base=reinterpret_cast<unsigned>(m_Buffer-sizeof (StringClass::_HEADER));
-		unsigned temp_base=reinterpret_cast<unsigned>(m_TempStrings+MAX_TEMP_BYTES*MAX_TEMP_STRING);
+		uintptr_t buffer_base=reinterpret_cast<uintptr_t>(m_Buffer-sizeof (StringClass::_HEADER));
+		uintptr_t temp_base=reinterpret_cast<uintptr_t>(m_TempStrings+MAX_TEMP_BYTES*MAX_TEMP_STRING);
 
 		if ((buffer_base>>11)==(temp_base>>11)) {
 			m_Buffer[0] = 0;
